% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operaciones.R
\name{media}
\alias{media}
\title{Media de valores de un numero o conjunto de numeros}
\usage{
media(x)
}
\arguments{
\item{x}{El primer operando}
}
\description{
Toma un numero o un conjunto de numeros y calcula su media
}

