% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operaciones.R
\name{raiz.cuadrada}
\alias{raiz.cuadrada}
\title{Raiz cuadrada de un valor numerico}
\usage{
raiz.cuadrada(x)
}
\arguments{
\item{x}{El primer operando}
}
\description{
Toma un valor numerico y calcula su raiz cuadrada
}

