% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operaciones.R
\name{resta.dos.numeros}
\alias{resta.dos.numeros}
\title{Resta de dos valores numericos}
\usage{
resta.dos.numeros(x, y)
}
\arguments{
\item{x}{El primer operando}

\item{y}{El segundo operando}
}
\description{
Toma dos valores y los resta
}

