% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operaciones.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma de dos valores numericos}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{El primer operando}

\item{y}{El segundo operando}
}
\description{
Toma dos valores y los suma
}

